# externalDB/swagger.yml
swagger: '2.0'
info:
  title: Your API
  version: 1.0.0
paths:
  /get_entity:
    get:
      summary: Get an entity by ID
      operationId: getEntityRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: id
          in: query
          description: ID of the entity to be retrieved
          required: true
          type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Entity not found
  /post_entity:
    post:
      summary: Post an entity
      operationId: postEntityRoute
      x-swagger-router-controller: app  # Module name
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: entity
          in: body
          required: true
          schema:
            $ref: '#/definitions/Entity'
      responses:
        '200':
          description: Successful response
  /update_entity:
    patch:  # Adding PATCH method
      summary: Update an entity
      operationId: patchEntityRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: entityId
          in: query
          required: true
          type: string
        - name: atttributesUpdated
          in: body
          description: Object containing the updated attributes and their values
          required: true
          schema:
            $ref: '#/definitions/Entity'
      responses:
        '200':
          description: Successful response  
        '404':
          description: Entity not found
  /delete_entity:
    delete:
      summary: Delete an entity by ID
      operationId: deleteEntityRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: id
          in: query
          description: ID of the entity to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Entity not found

definitions:
  Entity:
    type: object
    properties:
      id:
        type: string
      type:
        type: string
      version:
        type: string
    required:
      - id
      - type
      - version

  CrowdFlowObserved:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          dateObserved:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          peopleCount:
            type: object
            properties:
              type:
                type: string
              value:
                type: number
          congested:
            type: object
            properties:
              type:
                type: string
              value:
                type: boolean

  TrafficViolation:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          dateObserved:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          owner:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          description:
            type: object
            properties:
              type:
                type: string
              value:
                type: string

  TransportStation:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          name:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          category:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          dateLastReported:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          vehicleLastReported:
            type: object
            properties:
              type:
                type: string
              value:
                type: string

  Vehicle:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          fleetVehicleId:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          servicestatus:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          brandname:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          modelname:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          licenseplate:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          fuellevel:
            type: object
            properties:
              type:
                type: string
              value:
                type: number

  Location:
    type: object
    properties:
      type:
        type: string
      coordinates:
        type: array
        items:
          type: number